requirements.txt
flask>=3.0.0
anthropic>=0.25.0
openai-whisper>=20231117
pdfplumber>=0.10.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
reportlab>=4.0.0
